+++ spec/avromatic/model/messaging_serialization_spec.rb
@@ -170,6 +170,11 @@ describe Avromatic::Model::MessagingSerialization do
     end
   end
 
+  it_behaves_like "logical type encoding and decoding" do
+    let(:encoded_value) { instance.avro_message_value }
+    let(:decoded) { test_class.avro_message_decode(encoded_value) }
+  end
+
   context "custom types" do
     let(:schema_name) { 'test.named_type' }
     let(:test_class) do